:recycle: (admin) 리프레시 토큰 처리 및 인증 로직을 auth.ts와 lib/auth.ts로 통합, 불필요한 파일 삭제
