+++ apps/admin/src/components/auth/AuthInit.tsx
@@ -1,9 +1,8 @@
 import { useEffect } from 'react';
 
-import { authorization } from '@/model/auth/tokenStorage';
-import { useToken } from '@/model/auth/useToken';
-
 import { setHeaders } from '@/api/client';
+import { authorization } from '@/lib/auth';
+import { useToken } from '@/model/auth/useToken';
 
 /**
  * 앱 최상단에서 한 번 마운트. sessionStorage에 토큰이 있으면
